Moved classes to more appropriate files
